airflow-helper/requirements.txt
# REQUIREMENTS
# ------------------------------
# Contains any dependencies for running in CI pipeline and Dev Container.

pytest
ruff

# airflow libraries from composer setup 
apache-airflow==2.7.3
apache-airflow-providers-sendgrid
apache-airflow-providers-google
google-cloud-secret-manager
google-re2==1.0

# documentation as code
google-cloud-bigquery
google-cloud
lazydocs==0.4.8
mkdocs
mkdocstrings
mkdocstrings-python
pandas